# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DecisionDemoInput {
  /**
     * @minLength 12
     * @maxLength 1000
     */
  question: string;
}

export interface DecisionContext {
  location: string;
  coordinates: string;
  timeWindow: string;
  timezone: string;
  activity: string;
  vessel: string;
  extractedFrom: string;
}

export type WorkflowStageStatus = typeof WorkflowStageStatus[keyof typeof WorkflowStageStatus];


export const WorkflowStageStatus = {
  fixture: 'fixture',
  stale: 'stale',
  unavailable: 'unavailable',
  derived: 'derived',
  assumed: 'assumed',
} as const;

export interface WorkflowStage {
  id: string;
  label: string;
  status: WorkflowStageStatus;
  detail: string;
  /** @nullable */
  sourceRef: string | null;
  freshness: string;
}

export interface RiskDriver {
  label: string;
  signal: string;
  impact: string;
  sourceRef: string;
}

export type EvidenceItemType = typeof EvidenceItemType[keyof typeof EvidenceItemType];


export const EvidenceItemType = {
  derived: 'derived',
  assumed: 'assumed',
  unavailable: 'unavailable',
  fixture: 'fixture',
} as const;

export type EvidenceItemStatus = typeof EvidenceItemStatus[keyof typeof EvidenceItemStatus];


export const EvidenceItemStatus = {
  fixture: 'fixture',
  stale: 'stale',
  unavailable: 'unavailable',
  used: 'used',
} as const;

export interface EvidenceItem {
  id: string;
  claim: string;
  type: EvidenceItemType;
  source: string;
  status: EvidenceItemStatus;
  timestamp: string;
  operation: string;
  usedInFinding: boolean;
  detail: string;
}

export interface DecisionDemoResult {
  runId: string;
  scenarioLabel: string;
  runStatus: string;
  generatedAt: string;
  context: DecisionContext;
  stages: WorkflowStage[];
  finding: string;
  findingQualifier: string;
  confidence: string;
  confidenceReason: string;
  riskDrivers: RiskDriver[];
  assumptions: string[];
  limitations: string[];
  mapLabel: string;
  mapNote: string;
  evidence: EvidenceItem[];
}

export interface SavedDecision {
  id: number;
  question: string;
  createdAt: string;
  result: DecisionDemoResult;
}