# lib/api-zod/src/generated/types/savedDecision.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionDemoResult } from './decisionDemoResult';

export interface SavedDecision {
  id: number;
  question: string;
  createdAt: Date;
  result: DecisionDemoResult;
}